backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
deprecated==1.2.18
distro==1.9.0
executing==2.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.76.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pygithub==2.6.1
pygments==2.19.1
pyjwt==2.10.1
pynacl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyzmq==26.4.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
wrapt==1.17.2
psycopg2-binary==2.9.10
